fix stream aggregate function

+++ integration_pulsar_stream/src/main/java/ru/syntez/integration/pulsar/entities/ComposeDocument.java
@@ -14,7 +14,7 @@ import java.util.List;
 @JsonIgnoreProperties(ignoreUnknown = true)
 @Data
 public class ComposeDocument {
-    private int composeId;
+    private Long composeId;
     private String composeLabel;
     private List<OrderDocument> orderDocuments;
     private List<InvoiceDocument> invoiceDocuments;
